+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Context moved from http://schema.org to https://schema.org/ (#58)
 - By default, markup is added to noindex pages, a configuration setting is available for deactivation (#60)
 
+### Fixed
+- Custom doktypes greater than 199 are rendered in breadcrumb list
+
 ### Removed
 - Compatibility with TYPO3 v9 LTS (#41)
 - Compatibility with PHP 7.2 and PHP 7.3 (#42)